requirements.txt
BeautifulSoup==3.2.1
Collectfast==0.2.3
Django==1.8.2
Pillow==2.9.0
argparse==1.3.0
beautifulsoup4==4.4.0
boto==2.38.0
docopt==0.6.2
dj-database-url==0.3.0
dj-static==0.0.6
django-appconf==1.0.1
django-bootstrap3==6.1.0
django-ckeditor==4.4.8
django-classy-tags==0.6.2
git+https://h3@github.com/h3/django-colorfield.git   
django-compressor==1.5
django-cookie-law==1.0.4 
django-countries==3.3
django-grappelli==2.6.5
django-haystack==2.4.0
django-heroku-memcacheify==0.8
django-heroku-postgresify==0.4
django-modeltranslation==0.10
django-pylibmc==0.6.0
django-pylibmc-sasl==0.2.4
django-recaptcha==1.0.4
django-robots==1.1
django-storages==1.1.8
django-template-minifier==1.0
django-toolbelt==0.0.1
django-validated-file==2.0.1
elasticsearch==1.6.0 
gevent==1.0.2
grappelli-modeltranslation==0.1.3
greenlet==0.4.7
gunicorn==19.3.0
html5lib==0.99999
mandrill==1.0.57
newrelic==2.52.0.40
pylibmc==1.4.1
psycopg2==2.6.1
python-magic==0.4.6  
pytz==2015.4
requests==2.7.0
six==1.9.0 
slimit==0.8.1
sorl-thumbnail==12.2
static3==0.6.1
virtualenv==13.1.0
wsgiref==0.1.2

